got the userprofiles by interest working - updated schema, seed, interest query and controller
