+++ db/schema.sql
@@ -19,7 +19,6 @@ CREATE TABLE user_profiles (
 
 CREATE TABLE interests (
     interest_id SERIAL PRIMARY KEY,
-    userprofile_id INTEGER REFERENCES user_profiles(userprofile_id),
     name TEXT
 );
 
@@ -36,7 +35,6 @@ CREATE TABLE goals (
 );
 
 CREATE TABLE interest_connections (
-    interestconnections_id SERIAL PRIMARY KEY,
     userprofile_id INTEGER REFERENCES user_profiles(userprofile_id),
     interest_id INTEGER REFERENCES interests(interest_id)
 );
