+++ queries/interest.js
@@ -1,6 +1,5 @@
 const db = require("../db/dbConfig");
 
-
 const getInterests = async () => {
   try {
     const interests = await db.any("SELECT * FROM interests");
@@ -10,12 +9,12 @@ const getInterests = async () => {
   }
 };
 
-
-
-
 const getInterest = async (userprofile_id) => {
   try {
-    const interest = await db.any("SELECT * FROM interests WHERE userprofile_id=$1 ", [userprofile_id]);
+    const interest = await db.any(
+      "SELECT * FROM interests WHERE userprofile_id=$1 ",
+      [userprofile_id]
+    );
     return interest;
   } catch (error) {
     return error;
@@ -38,15 +37,14 @@ const getInterest = async (userprofile_id) => {
 //     throw error;
 //   }
 // };
-const getUserProfilesByInterest = async (interestconnections_id,interest_id) => {
+
+const getUserProfilesByInterest = async (interest_id) => {
   try {
-    const interests = await db.one(
-      "SELECT * FROM interest_connections WHERE interestconnections_id=$1 AND interest_id=$2"
-      
-      
-      ,
-      [interestconnections_id,interest_id]
+    const interests = await db.any(
+      "SELECT * FROM interest_connections JOIN user_profiles ON interest_connections.userprofile_id = user_profiles.userprofile_id WHERE interest_id=$1",
+      [interest_id]
     );
+    console.log(interests);
     return interests;
   } catch (error) {
     throw error;
@@ -56,5 +54,5 @@ const getUserProfilesByInterest = async (interestconnections_id,interest_id) => 
 module.exports = {
   getInterests,
   getInterest,
-  getUserProfilesByInterest
+  getUserProfilesByInterest,
 };
